third commit

+++ ElevanceProject/src/main/java/com/demo/mongodb/controller/TraineeController.java
@@ -3,10 +3,19 @@
  */
 package com.demo.mongodb.controller;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.LocalDate;
+import java.time.format.DateTimeParseException;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,8 +26,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.demo.mongodb.entity.Trainee;
+import com.demo.mongodb.entity.TraineeDAO;
+import com.demo.mongodb.service.TraineeResponse;
 import com.demo.mongodb.service.TraineeService;
-
+import static com.demo.mongodb.util.InputDataValidation.*;
+import static com.demo.mongodb.util.Constants.*;
 /**
  * @author nikhilgupta2
  *
@@ -26,83 +38,219 @@ import com.demo.mongodb.service.TraineeService;
 @RestController
 public class TraineeController {
 	
+	Logger logger = LoggerFactory.getLogger(TraineeService.class);
+
+	
 	@Autowired
 	private TraineeService service;
 	
 	@GetMapping("/trainee")
-	private List<Trainee> getTrainees() {
+	private ResponseEntity<TraineeResponse> getTrainees() {
 		return service.getAllTrainees();
 	}
 	
 	@GetMapping("/trainee/{tid}")
-	private Optional<Trainee> getTraineeById(@PathVariable int tid) {
+	private ResponseEntity<Optional<Trainee>> getTraineeById(@PathVariable long tid) {
 
-		return service.getTraineebyID(tid);
+		try {
+			if(isNegativeId(tid))
+				throw new IllegalArgumentException(NEGVAL);
+			
+			return service.getTraineebyID(tid);
+			}
+		catch(IllegalArgumentException e)
+			{
+				logger.error(e.getMessage());
+				return new ResponseEntity<>(null,HttpStatus.BAD_REQUEST);
+			} 
 	}
 	
 	@GetMapping("/traineeWithPages")
-	private List<Trainee> getTraineeWithPages(@RequestParam int pn, @RequestParam int ps) {
+	private ResponseEntity<TraineeResponse> getTraineeWithPages(
+			@RequestParam(value="pageno", required=true, defaultValue ="1") int pn
+	, @RequestParam(value="pagesiz", required=true,defaultValue = "1") int ps
+	, @RequestParam(value="dir", required=true,defaultValue = "1") int dir
+	, @RequestParam(value="sortby", required=true,defaultValue = "id") String s) {
+		
+		TraineeResponse response=new TraineeResponse();
 
-		return service.getAllTraineesPages(pn,ps);
+		try {
+			if(isNegativeVal(pn, ps))
+				throw new IllegalArgumentException(NEGVAL);
+			if(dirnValidation(dir))
+				throw new IllegalArgumentException(DIRERROR);
+			
+			return service.getAllTraineesPages(pn,ps,dir,s);
+		}
+		catch(IllegalArgumentException e)
+			{
+				logger.error(e.getMessage());
+				return new ResponseEntity<>(response,HttpStatus.BAD_REQUEST);
+		}
 	}
 	
 	//Get trainees with sorting. 1 for asc, -1 for desc.
-	@GetMapping("/traineeWithSort/{dir}")
-	private List<Trainee> getTraineeWithSorting(@PathVariable int dir) {
+	@GetMapping("/traineeWithSort/{dir}/{by}")
+	private ResponseEntity<TraineeResponse> getTraineeWithSorting(@PathVariable int dir,@PathVariable String by) {
 
-		return service.getSortedTrainees(dir);
+		TraineeResponse response=new TraineeResponse();
+
+		try {
+			if(dirnValidation(dir))
+				throw new IllegalArgumentException(DIRERROR);
+			return service.getSortedTrainees(dir,by);
+		}
+		catch(IllegalArgumentException e)
+			{
+				logger.error(e.getMessage());
+				return new ResponseEntity<>(response,HttpStatus.BAD_REQUEST);
+		}
 	}
 	
 	// Get trainees with ids greater than x
 	@GetMapping("/traineeGreaterThan/{gt}")
-	private List<Trainee> getTraineeGreater(@PathVariable int gt) {
+	private ResponseEntity<TraineeResponse> getTraineeGreater(@PathVariable int gt) {
 
-		return service.getTraineesGreaterThan(gt);
+		TraineeResponse response=new TraineeResponse();
+
+		try {
+			if(isNegativeVal(gt))
+				throw new IllegalArgumentException(NEGVAL);
+			return service.getTraineesGreaterThan(gt);
+		}
+		catch(IllegalArgumentException e)
+			{
+				logger.error(e.getMessage());
+				return new ResponseEntity<>(response,HttpStatus.BAD_REQUEST);
+			}
 	}
 
 	// Get trainees by Name
 	@GetMapping("/traineeByName/{nam}")
-	private List<Trainee> getTraineeThroughName(@PathVariable String nam) {
+	private ResponseEntity<TraineeResponse> getTraineeThroughName(@PathVariable String nam) {
 
-		return service.getTraineeByName(nam);
+		TraineeResponse response=new TraineeResponse();
+		
+		try {
+			if(isNotValidString(nam))
+				throw new IllegalArgumentException("name " + REQUIRED);
+			return service.getTraineeByName(nam);
+		}
+		catch (IllegalArgumentException e) {
+			logger.error(e.getMessage());
+			return new ResponseEntity<>(response,HttpStatus.BAD_REQUEST);
+		}
+		
 	}
 	
 	// Get trainees by Name display only Location
-		@GetMapping("/traineeByNameDispLocation/{nam}")
-		private List<Trainee> getTraineeThroughNameDispLocation(@PathVariable String nam) {
+	@GetMapping("/traineeByNameDispLocation/{nam}")
+	private ResponseEntity<TraineeResponse> getTraineeThroughNameDispLocation(@PathVariable String nam) {
 
+		TraineeResponse response=new TraineeResponse();
+
+		try {
+			if(isNotValidString(nam))
+				throw new IllegalArgumentException("name " + REQUIRED);
 			return service.getTraineeByNameDispLoc(nam);
 		}
+		catch(IllegalArgumentException e)
+			{
+				logger.error(e.getMessage());
+				return new ResponseEntity<>(response,HttpStatus.BAD_REQUEST);
+			}
+	}
 
 	// Get trainees by Name and Location
 	@GetMapping("/traineeByNameAndLoc/{nam}/{loc}")
-	private List<Trainee> getTraineeThroughNameLocation(@PathVariable String nam,@PathVariable String loc) {
+	private ResponseEntity<TraineeResponse> getTraineeThroughNameLocation(@PathVariable String nam,@PathVariable String loc) {
+		TraineeResponse response=new TraineeResponse();
 
-		return service.getTraineeByNameAndLocation(nam,loc);
+		try {
+			if((isNotValidString(nam)||(isNotValidString(loc))))
+				throw new IllegalArgumentException("name and location" + REQUIRED);
+			return service.getTraineeByNameAndLocation(nam,loc);
+		}
+		catch (IllegalArgumentException e) {
+			logger.error(e.getMessage());
+			return new ResponseEntity<>(response,HttpStatus.BAD_REQUEST);
+		}
+		
 	}
 	
 	// Get trainees by Name Like
 	@GetMapping("/traineeByNameRegex/{expr}")
-	private List<Trainee> getTraineewithNameregex(@PathVariable String expr) {
+	private ResponseEntity<TraineeResponse> getTraineewithNameregex(@PathVariable String expr) {
 
 		return service.getTraineeByNameContaining(expr);
 	}
 	
 	// Get trainee count by Name
-		@GetMapping("/traineeCountByName/{nam}")
-		private int getTraineeCountThroughName(@PathVariable String nam) {
+	@GetMapping("/traineeCountByName/{nam}")
+	private int getTraineeCountThroughName(@PathVariable String nam) {
 
+		try {
+			if(isNotValidString(nam))
+				throw new IllegalArgumentException("name " + REQUIRED);
 			return service.getTraineeCountByName(nam);
+		} 
+		catch (IllegalArgumentException e) {
+			logger.error(e.getMessage());
+			return 0;
 		}
+		
+	}
 	
 	@PostMapping("/trainee")
-	private Trainee saveTrainee(@RequestBody Trainee t ) {
-		return service.createTrainee(t);
+	private ResponseEntity<Trainee> saveTrainee(@RequestBody TraineeDAO t ) throws ParseException {
+		
+	
+		try {
+			if(isNotValidDate(t.getJoinDate(), t.getTestDate()))
+				throw new IllegalArgumentException(INVALIDATE);
+	
+			if((isNotValidString(t.getName())||(isNotValidString(t.getLocation()))))
+				throw new IllegalArgumentException("name and location" + REQUIRED);
+			
+			if(isNegativeId(t.getId()))
+				throw new IllegalArgumentException(NEGVAL);
+			
+			//SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy");
+			LocalDate jd = LocalDate.parse(t.getJoinDate());
+			LocalDate td = LocalDate.parse(t.getTestDate());
+
+			return service.createTrainee(new Trainee(t.getId(),t.getName(),t.getLocation(),
+					jd,td));
+		}
+		catch(IllegalArgumentException|DateTimeParseException e){
+			logger.error(e.getMessage());
+			return new ResponseEntity<>(null, HttpStatus.EXPECTATION_FAILED);
+		}
 	}
 	
 	@PutMapping("/trainee")
-	private Trainee updateTrainee(@RequestBody Trainee t ) {
-		return service.updateTrainee(t);
+	private ResponseEntity<Trainee> updateTrainee(@RequestBody TraineeDAO t ) {
+		
+		try {
+			if(isNotValidDate(t.getJoinDate(), t.getTestDate()))
+				throw new IllegalArgumentException(INVALIDATE);
+			if((isNotValidString(t.getName())||(isNotValidString(t.getLocation()))))
+				throw new IllegalArgumentException("name and location" + REQUIRED);
+			if(isNegativeId(t.getId()))
+				throw new IllegalArgumentException(NEGVAL);
+	
+			//SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy");
+			LocalDate jd = LocalDate.parse(t.getJoinDate());
+			LocalDate td = LocalDate.parse(t.getTestDate());
+
+			return service.updateTrainee(new Trainee(t.getId(),t.getName(),t.getLocation(),
+					jd,td));
+		}
+		catch(IllegalArgumentException|DateTimeParseException e){
+			logger.error(e.getMessage());
+			return new ResponseEntity<>(null,HttpStatus.EXPECTATION_FAILED);		
+
+			}
 	}	
 	
 	/*
@@ -113,9 +261,17 @@ public class TraineeController {
 	
 	
 	@DeleteMapping("/trainee/{tid}")
-	private String deleteTrainee(@PathVariable int tid) {
-		return service.deleteTraineebyID(tid);
+	private void  deleteTrainee(@PathVariable long tid) {
 		
+		try {
+			if(isNegativeId(tid))
+				throw new IllegalArgumentException(NEGVAL);
+			service.deleteTraineebyID(tid);
+		}
+		catch(IllegalArgumentException e)
+		{
+			logger.error(e.getMessage());
+		}
 	}
 
 }
